requirements.txt
torch 
torchvision 
transformers 
pillow 
matplotlib 
openai-clip